// src/types/LaunchItem.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { InputPipeline } from "./InputPipeline";
import type { ParamPreset } from "./ParamPreset";

export type LaunchItem = { id: string, name: string, type: string, url: string | null, executablePath: string | null, args: string | null, cwd: string | null, envVariables: string | null, runAsAdmin: boolean | null, inTerminal: boolean | null, isDir: boolean | null, iconUrl: string | null, shortcut: string | null, inputPipeline: InputPipeline | null, paramPresets: Array<ParamPreset> | null, multiParamEnabled: boolean | null, categoryId: string, columnId: string, };