// src/types/SettingsConfig.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Tab } from "./Tab";

export type SettingsConfig = { columns: number, summonShortcut: string, summonMouseShortcut: string | null, dockPosition: string, leftTabs: Array<Tab>, topTabs: { [key in string]?: Array<Tab> }, };